flask-dashboard/requirements.txt
Flask==3.0.3
boto3==1.35.0
botocore==1.35.0
plotly==5.24.1
python-dotenv==1.0.1

